assignment/tools/requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==68.2.2
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
